split portrait placement file
