+++ data/data_8107224.s
@@ -1,22 +1,18 @@
 .section .rodata
 
 
-
+.string "pksdir0\0" @Groudon
+.string "pksdir0\0" @Ray
+.string "pksdir0\0" @Entei
+.string "pksdir0\0" @Raikou
+.string "pksdir0\0" @Suicune
+.string "pksdir0\0" @HoOh
+.string "pksdir0\0" @Latios
 .string "pksdir0\0"
 .string "pksdir0\0"
 .string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-@ ???
+.string "pksdir0\0" @ 
+@ ??? Jirachi
 .string "pksdir0\0"
 
 .global gUnknown_81074FC
@@ -24,34 +20,18 @@ gUnknown_81074FC: @ 81074FC
 .byte 0x4e, 0x4f, 0x50, 0x51, 0x4a, 0x49, 0x3e, 0x38
 
 @ ???
-.string "pksdir0\0"
+.string "pksdir0\0" @ lugia
 @ ???
-.string "pksdir0\0"
+.string "pksdir0\0" @kyorge
 @ ???
-.string "pksdir0\0"
+.string "pksdir0\0" @ deoxys
 @ ???
-.string "pksdir0\0"
+.string "pksdir0\0" @celebi
 @ ???
-.string "pksdir0\0"
+.string "pksdir0\0" @unk
 @ ???
-.string "pksdir0\0"
+.string "pksdir0\0" @ medicham
 @ ???
-.string "pksdir0\0"
-@ ???
-.string "pksdir0\0"
-
-.global gUnknown_8107544
-gUnknown_8107544: @ 8107544
-.byte 0x02, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x05, 0x00, 0x00, 0x00, 0x00, 0x00
-.byte 0x02, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x17, 0x00, 0x08, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x07, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x00, 0x08, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x0d, 0x00, 0x05, 0x00, 0x01, 0x00, 0x00, 0x00, 0x02, 0x00, 0x08, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x17, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x08, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x12, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00
-.byte 0x02, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x17, 0x00, 0x02, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x07, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x0d, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x02, 0x00, 0x02, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x17, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00
-.byte 0x12, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00
+.string "pksdir0\0" @ smeargle
 
 .align 2,0
