+++ src/dungeon_cutscenes.c
@@ -72,7 +72,7 @@ extern void sub_8086A3C(Entity *pokemon);
 extern void SetupBossFightHP(Entity *pokemon, s32 newHP, u16 songIndex);
 extern u8 sub_8086AE4(s16 _index);
 
-void sub_8085F44(s32);
+
 void sub_808BBA8(Entity * );
 void DeoxysScreenFlash(void);
 void sub_8086A54(Entity *);
@@ -142,592 +142,17 @@ void SkarmoryEntry(Entity *);
 
 // FILE SPLIT
 
-void sub_80877E8(void)
-{
-  Entity * leaderEntity;
-  Entity * moltresEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  moltresEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MOLTRES);
-  DungeonStartNewBGM(MUS_THE_MOUNTAIN_OF_FIRE);
-  sub_8085374();
-  sub_80854D4();
-  sub_8085930(DIRECTION_WEST);
-  sub_80855E4(sub_8086A3C);
-  SetFacingDirection(moltresEntity, DIRECTION_SOUTH);
-  sub_8086A3C(moltresEntity);
-  sub_8085860(leaderEntity->pos.x + -3,leaderEntity->pos.y + -2);
-  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2], MONSTER_MOLTRES);
-}
-
-void sub_8087848(void)
-{
-  Entity *leaderEntity;
-  Entity *moltresEntity;
-
-  u32 XPos;
-  s32 YPos;
-
-  leaderEntity = xxx_call_GetLeader();
-  moltresEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MOLTRES);
-  DungeonStartNewBGM(MUS_THE_MOUNTAIN_OF_FIRE);
-  sub_80854D4();
-  sub_8085930(DIRECTION_WEST);
-  sub_80855E4(sub_8086A3C);
-  if (sub_8086AE4(MONSTER_MOLTRES)) {
-    sub_8068FE0(moltresEntity,0x21c,0);
-  }
-  else {
-    sub_8072008(moltresEntity,moltresEntity,gMoltresConfigLevel,0,0);
-    SetFacingDirection(moltresEntity, DIRECTION_SOUTH);
-    sub_8086A3C(moltresEntity);
-  }
-  sub_8085860(leaderEntity->pos.x - 5, leaderEntity->pos.y);
-  XPos = GetCameraXPos();
-  YPos = GetCameraYPos();
-  sub_803F878(XPos, YPos + -0x1000);
-  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2],MONSTER_MOLTRES);
-}
-
-void sub_80878F4(u8 param_1, u8 param_2)
-{
-  if ((param_2 == 0xA || param_2 == 0xB || param_2 == 0xC) && (param_1 == 0x9)) {
-    sub_8097FA8(0x7);
-    gDungeon->unk2 = 1;
-  }
-}
-
-void MoltresPreFightDialogue(void)
-{
-  Entity *leaderEntity;
-  Entity *partnerEntity;
-  Entity *moltresEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  partnerEntity = GetPartnerEntity();
-  moltresEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MOLTRES);
-  MoltresScreenFlash1(0xc,5);
-  sub_8086500();
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(&gMoltresPreFightDialogue_1);
-  sub_803E708(10,0x46);
-  sub_80869E4(partnerEntity,4,1,DIRECTION_NORTHWEST);
-  sub_803E708(4,0x46);
-  SpriteLookAroundEffect(partnerEntity);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(gMoltresPreFightDialogue_2);
-  sub_803E708(10,0x46);
-  MoltresScreenFlash1(0xc,5);
-  MoltresScreenFlash2(9,5);
-  DisplayDungeonDialogue(gMoltresPreFightDialogue_3);
-  sub_803E708(10,0x46);
-  sub_80869E4(partnerEntity,4,1,DIRECTION_EAST);
-  sub_80869E4(leaderEntity,4,2,DIRECTION_WEST);
-  DisplayDungeonDialogue(gMoltresPreFightDialogue_4);
-  sub_803E708(10,0x46);
-  sub_80869E4(partnerEntity,4,2,DIRECTION_WEST);
-  sub_8086690();
-  DisplayDungeonDialogue(&gMoltresPreFightDialogue_5);
-  sub_803E708(10,0x46);
-  sub_8085930(DIRECTION_NORTHWEST);
-  sub_803E708(4,0x46);
-  sub_8085930(DIRECTION_NORTH);
-  PlaySoundEffect(0x2f2);
-  DungeonStopBGM();
-  MoltresScreenDarken();
-  DisplayDungeonDialogue(gMoltresPreFightDialogue_6);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(&gMoltresPreFightDialogue_7);
-  sub_803E708(10,0x46);
-  PlaySoundEffect(0x1cf);
-  sub_80856E0(partnerEntity, DIRECTION_NORTH);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(gMoltresPreFightDialogue_8);
-  sub_803E708(10,0x46);
-  MoltresScreenFlash3();
-  PlaySoundEffect(0x1da);
-  DisplayDungeonDialogue(&gMoltresPreFightDialogue_9);
-  sub_803E708(10,0x46);
-  MoltresScreenFlash1(9,5);
-  MoltresScreenFlash2(0xb,5);
-  DisplayDungeonDialogue(&gMoltresPreFightDialogue_10);
-  sub_803E708(10,0x46);
-  MoltresScreenFlash1(0xc,5);
-  MoltresScreenFlash2(9,5);
-  MoltresDropInEffect(moltresEntity);
-  sub_808663C();
-  MoltresScreenFlash3();
-  DisplayDungeonDialogue(&gMoltresPreFightDialogue_11);
-  sub_803E708(10,0x46);
-  MoltresScreenFlash2(0xb,6);
-  DisplayDungeonDialogue(&gMoltresPreFightDialogue_12);
-  sub_803E708(10,0x46);
-  SetupBossFightHP(moltresEntity,400,MUS_BOSS_BATTLE);
-  ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
-}
-
-void MoltresReFightDialogue(void)
-{
-  Entity *leaderEntity;
-  Entity *partnerEntity;
-  Entity *moltresEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  partnerEntity = GetPartnerEntity();
-  moltresEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MOLTRES);
-  MoltresScreenFlash1(10,5);
-  sub_8086500();
-  sub_803E708(10,0x46);
-  SpriteLookAroundEffect(partnerEntity);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(&gMoltresReFightDialogue_1);
-  sub_803E708(10,0x46);
-  sub_80869E4(partnerEntity,4,1,DIRECTION_EAST);
-  sub_80869E4(leaderEntity,4,2,DIRECTION_WEST);
-  DisplayDungeonDialogue(gMoltresReFightDialogue_2);
-  sub_803E708(10,0x46);
-  sub_80869E4(partnerEntity,4,2,DIRECTION_WEST);
-  sub_80869E4(leaderEntity,4,2,DIRECTION_WEST);
-  sub_8086690();
-  MoltresScreenDarken();
-  sub_803E708(10,0x46);
-  sub_8085930(DIRECTION_NORTHWEST);
-  sub_803E708(4,0x46);
-  sub_8085930(DIRECTION_NORTH);
-  MoltresDropInEffect(moltresEntity);
-  sub_808663C();
-  MoltresScreenFlash3();
-  DisplayDungeonDialogue(&gMoltresReFightDialogue_3);
-  sub_803E708(10,0x46);
-  MoltresScreenFlash2(0xb,6);
-  DisplayDungeonDialogue(gMoltresReFightDialogue_4);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(&gMoltresReFightDialogue_5);
-  sub_803E708(10,0x46);
-  SetupBossFightHP(moltresEntity,400,MUS_BOSS_BATTLE);
-  ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
-}
-
-void MoltresPostStoryPreFightDialogue(void)
-{
-  Entity * leaderEntity;
-  Entity * moltresEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  moltresEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MOLTRES);
-  MoltresScreenFlash1(0xc,5);
-  sub_808654C();
-  if (sub_8086AE4(MONSTER_MOLTRES)) {
-    sub_8085930(DIRECTION_NORTHWEST);
-    sub_803E708(4,0x46);
-    sub_8085930(DIRECTION_NORTH);
-    sub_803E708(4,0x46);
-    sub_803E708(10,0x46);
-    DisplayDungeonDialogue(&gUnknown_8100D3C);
-    sub_803E708(10,0x46);
-    gDungeon->unk2 = 1;
-  }
-  else {
-    DisplayDungeonDialogue(&gMoltresPostStoryPreFightDialogue_1);
-    sub_8085930(DIRECTION_NORTHWEST);
-    sub_803E708(4,0x46);
-    sub_8085930(DIRECTION_NORTH);
-    MoltresDropInEffect(moltresEntity);
-    DisplayDungeonDialogue(&gMoltresPostStoryPreFightDialogue_2);
-    MoltresScreenFlash1(0xc,5);
-    MoltresScreenFlash2(9,5);
-    DisplayDungeonDialogue(&gMoltresPostStoryPreFightDialogue_3);
-    MoltresScreenFlash3();
-    DisplayDungeonDialogue(&gMoltresPostStoryPreFightDialogue_4);
-    MoltresScreenFlash2(0xb,6);
-    DisplayDungeonDialogue(&gMoltresPostStoryPreFightDialogue_5);
-    sub_803E708(10,0x46);
-    SetupBossFightHP(moltresEntity,400,MUS_BOSS_BATTLE);
-    ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
-  }
-}
-
-void MoltresDropInEffect(Entity * moltresEntity)
-{
-  s32 iVar1;
-
-  GetEntInfo(moltresEntity)->unk15C = 1;
-  GetEntInfo(moltresEntity)->unk15E = 0;
-  GetEntInfo(moltresEntity)->unk174 = IntToF248_2(200);
-  PlaySoundEffect(0x1f8);
-  for(iVar1 = 200; iVar1 >= 0; iVar1 -= 5)
-  {
-    GetEntInfo(moltresEntity)->unk174 = IntToF248_2(iVar1);
-    DungeonRunFrameActions(0x46);
-  }
-}
-
-void MoltresScreenFlash1(s32 xArg, s32 yArg)
-{
-    s32 x, y;
-    DungeonPos newPos;
-
-    x = X_POS_TO_PIXELPOS(xArg);
-    newPos.x = x / 256;
-
-    y = Y_POS_TO_PIXELPOS(yArg);
-    newPos.y = y / 256;
-
-    sub_8085EC8(370,16,0,&newPos,0);
-}
-
-void MoltresScreenFlash2(s32 xArg, s32 yArg)
-{
-    s32 r6;
-    s32 i, x, y;
-    DungeonPos newPos;
-
-    x = X_POS_TO_PIXELPOS(xArg);
-    newPos.x = x / 256;
-
-    y = Y_POS_TO_PIXELPOS(yArg);
-    newPos.y = y / 256;
-
-    r6 = sub_8085EC8(370,22,0,&newPos,0);
-    PlaySoundEffect(0x1EC);
-
-    for (i = 250; i >= 20; i -= 10) {
-        SetDungeonBGColorRGB(i, 0, 0, 1, 1);
-        DungeonRunFrameActions(0x46);
-    }
-    sub_803E708(0xA, 0x46);
-
-    for (i = 250; i >= 0; i -= 10) {
-        SetDungeonBGColorRGB(i, 0, 0, 1, 1);
-        DungeonRunFrameActions(0x46);
-    }
-    sub_803E708(0xA, 0x46);
-
-    sub_8085EB0();
-    sub_8085F44(r6);
-}
-
-void MoltresScreenFlash3(void)
-{
-  int iVar1;
-
-  PlaySoundEffect(0x1ed);
-  iVar1 = 0xfa;
-  for(iVar1 = 250; iVar1 > 9; iVar1 -= 10)
-  {
-    SetDungeonBGColorRGB(iVar1,iVar1,iVar1 / 2,1,1);
-    DungeonRunFrameActions(0x46);
-  }
-  sub_803E708(0x1e,0x46);
-  PlaySoundEffect(0x1ed);
-  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 10)
-  {
-    SetDungeonBGColorRGB(iVar1,iVar1,iVar1 / 2,1,1);
-    DungeonRunFrameActions(0x46);
-  }
-  sub_803E708(0x1e,0x46);
-  sub_8085EB0();
-}
-
-void MoltresScreenDarken(void)
-{
-  s32 iVar1;
-
-  for(iVar1 = 0; iVar1 >= -80; iVar1 -= 8)
-  {
-    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,0,0);
-    DungeonRunFrameActions(0x46);
-  }
-}
 
 
-void sub_8087F54(void)
-{
-  Entity  *leaderEntity;
-  Entity *articunoEntity;
+// SPLIT
 
-  leaderEntity = xxx_call_GetLeader();
-  articunoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_ARTICUNO);
-  DungeonStartNewBGM(MUS_IN_THE_DEPTHS_OF_THE_PIT);
-  sub_8085374();
-  sub_80854D4();
-  sub_8085930(DIRECTION_NORTH);
-  sub_80855E4(sub_8086A3C);
-  SetFacingDirection(articunoEntity, DIRECTION_SOUTH);
-  sub_8086A3C(articunoEntity);
-  sub_8085860(leaderEntity->pos.x,leaderEntity->pos.y + -3);
-  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2], MONSTER_ARTICUNO);
-}
 
-void sub_8087FB4(void)
-{
-  Entity  *articunoEntity;
 
-  articunoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_ARTICUNO);
-  DungeonStartNewBGM(MUS_IN_THE_DEPTHS_OF_THE_PIT);
-  sub_8085374();
-  sub_80854D4();
-  sub_8085930(DIRECTION_NORTH);
-  SetFacingDirection(articunoEntity, DIRECTION_SOUTH);
-  sub_806CDD4(articunoEntity, 0xF, DIRECTION_SOUTH);
-  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2], MONSTER_ARTICUNO);
-}
+// File Split
 
-void sub_8087FF8(void)
-{
-  Entity  *leaderEntity;
-  Entity *articunoEntity;
 
-  leaderEntity = xxx_call_GetLeader();
-  articunoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_ARTICUNO);
-  sub_8085374();
-  sub_80854D4();
-  sub_8085930(DIRECTION_NORTH);
-  sub_80855E4(sub_8086A3C);
-  if (sub_8086AE4(MONSTER_ARTICUNO)) {
-    sub_8068FE0(articunoEntity,0x21c,0);
-  }
-  else {
-    sub_8072008(articunoEntity,articunoEntity,gArticunoConfigLevel,0,0);
-    SetFacingDirection(articunoEntity, DIRECTION_SOUTH);
-    sub_8086A3C(articunoEntity);
-  }
-  sub_8085860(leaderEntity->pos.x,leaderEntity->pos.y + -3);
-  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2], MONSTER_ARTICUNO);
-}
 
-void sub_8088088(u8 param_1, u8 param_2)
-{
-  if ((param_2 == 0xD || param_2 == 0xE || param_2 == 0xF) && (param_1 == 0xD)) {
-    sub_8097FA8(0x9);
-    gDungeon->unk2 = 1;
-  }
-}
-
-void ArticunoPreFightDialogue(void)
-{
-  Entity *leaderEntity;
-  Entity *partnerEntity;
-  Entity *articunoEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  partnerEntity = GetPartnerEntity();
-  articunoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_ARTICUNO);
-  sub_8086448();
-  SpriteLookAroundEffect(partnerEntity);
-  DisplayDungeonDialogue(gArticunoPreFightDialogue_1);
-  sub_803E708(10,0x46);
-  sub_8086598();
-  DungeonStopBGM();
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_2);
-  sub_803E708(10,0x46);
-  sub_80869E4(partnerEntity,4,1,DIRECTION_EAST);
-  sub_80869E4(leaderEntity,4,2,DIRECTION_WEST);
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_3);
-  sub_803E708(10,0x46);
-  SetFacingDirection(leaderEntity, DIRECTION_WEST);
-  sub_803E708(4,0x46);
-  SetFacingDirection(leaderEntity, DIRECTION_NORTHWEST);
-  sub_803E708(4,0x46);
-  SetFacingDirection(leaderEntity, DIRECTION_WEST);
-  sub_803E708(4,0x46);
-  SetFacingDirection(leaderEntity, DIRECTION_SOUTHWEST);
-  sub_803E708(4,0x46);
-  SetFacingDirection(leaderEntity, DIRECTION_WEST);
-  sub_803E708(4,0x46);
-  SetFacingDirection(leaderEntity, DIRECTION_NORTHWEST);
-  sub_803E708(4,0x46);
-  SetFacingDirection(leaderEntity, DIRECTION_WEST);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(gArticunoPreFightDialogue_4);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_5);
-  sub_803E708(10,0x46);
-  sub_80869E4(leaderEntity,4,1,DIRECTION_NORTH);
-  sub_80869E4(partnerEntity,4,2,DIRECTION_NORTH);
-  PlaySoundEffect(0x1d5);
-  SpriteLookAroundEffect(partnerEntity);
-  PlaySoundEffect(0x1d5);
-  SpriteLookAroundEffect(leaderEntity);
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_6);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_7);
-  sub_803E708(10,0x46);
-  PlaySoundEffect(0); // snow flurry
-  gDungeon->weather.unkE265 = WEATHER_SNOW;
-  sub_807EAA0(0,1);
-  PlaySoundEffect(0x1ee); // flash effect?
-  sub_8088574();
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_8);
-  sub_803E708(10,0x46);
-  ArticunoScreenFlash();
-  sub_8086738();
-  PlaySoundEffect(0x1da);
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_9);
-  sub_803E708(10,0x46);
-  PlaySoundEffect(0x1ea);
-  sub_8088484(articunoEntity);
-  sub_80865E8();
-  ArticunoScreenFlash();
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_10);
-  sub_806CDD4(articunoEntity,0xf,DIRECTION_SOUTH);
-  ArticunoScreenFlash();
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_11);
-  sub_806CDD4(articunoEntity,0x10,DIRECTION_SOUTH);
-  ArticunoScreenFlash();
-  DisplayDungeonDialogue(&gArticunoPreFightDialogue_12);
-  sub_806CDD4(articunoEntity,0xf,DIRECTION_SOUTH);
-  sub_803E708(0x10,0x46);
-  SetupBossFightHP(articunoEntity,0x1c2,MUS_BOSS_BATTLE);
-  ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
-}
-
-void ArticunoReFightDialogue(void)
-{
-  Entity *leaderEntity;
-  Entity *articunoEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  articunoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_ARTICUNO);
-  gDungeon->weather.unkE265 = WEATHER_SNOW;
-  sub_807EAA0(0,1);
-  sub_803E708(10,0x46);
-  DisplayDungeonDialogue(&gArticunoReFightDialogue_1);
-  sub_806CDD4(articunoEntity,0x10,DIRECTION_SOUTH);
-  ArticunoScreenFlash();
-  DisplayDungeonDialogue(&gArticunoReFightDialogue_2);
-  ArticunoScreenFlash();
-  DisplayDungeonDialogue(&gArticunoReFightDialogue_3);
-  sub_806CDD4(articunoEntity,0xf,DIRECTION_SOUTH);
-  sub_803E708(0x28,0x46);
-  SetupBossFightHP(articunoEntity,0x1c2,MUS_BOSS_BATTLE);
-  ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
-}
-
-void ArticunoPostStoryPreFightDialogue(void)
-{
-  Entity *leaderEntity;
-  Entity *articunoEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  articunoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_ARTICUNO);
-  sub_8086448();
-  if (sub_8086AE4(MONSTER_ARTICUNO)) {
-    sub_80866C4(&gUnknown_8102A9C);
-  }
-  else {
-    gDungeon->weather.unkE265 = WEATHER_SNOW;
-    sub_807EAA0(0,1);
-    ArticunoScreenFlash();
-    sub_8086738();
-    PlaySoundEffect(0x1da);
-    DisplayDungeonDialogue(&gArticunoPostStoryPreFightDialogue_1);
-    sub_803E708(10,0x46);
-    sub_8088484(articunoEntity);
-    ArticunoScreenFlash();
-    DisplayDungeonDialogue(&gArticunoPostStoryPreFightDialogue_2);
-    sub_806CDD4(articunoEntity,0xf,DIRECTION_SOUTH);
-    ArticunoScreenFlash();
-    DisplayDungeonDialogue(&gArticunoPostStoryPreFightDialogue_3);
-    sub_806CDD4(articunoEntity,0x10,DIRECTION_SOUTH);
-    ArticunoScreenFlash();
-    DisplayDungeonDialogue(&gArticunoPostStoryPreFightDialogue_4);
-    ArticunoScreenFlash();
-    DisplayDungeonDialogue(&gArticunoPostStoryPreFightDialogue_5);
-    sub_806CDD4(articunoEntity,0xf,DIRECTION_SOUTH);
-    sub_803E708(0x28,0x46);
-    SetupBossFightHP(articunoEntity,0x1c2,MUS_BOSS_BATTLE);
-    ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
-  }
-}
-
-void sub_8088484(Entity *param_1)
-{
-  int iVar1;
-
-  GetEntInfo(param_1)->unk15C = 1;
-  GetEntInfo(param_1)->unk15E = 0;
-  PlaySoundEffect(0x1ea);
-  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 5)
-  {
-    GetEntInfo(param_1)->unk174 = IntToF248_2(iVar1);
-    SetDungeonBGColorRGB(iVar1,iVar1,iVar1 / 2,1,0);
-    DungeonRunFrameActions(0x46);
-  }
-}
-
-void ArticunoScreenFlash(void)
-{
-  int iVar1;
-
-  PlaySoundEffect(0x2ad);
-  for(iVar1 = 250; iVar1 > 0x95; iVar1 -= 10)
-  {
-    SetDungeonBGColorRGB(iVar1 / 2,iVar1,iVar1,1,1);
-    DungeonRunFrameActions(0x46);
-  }
-  sub_803E708(4,0x46);
-  for(iVar1 = 250; iVar1 > 199; iVar1 -= 10)
-  {
-    SetDungeonBGColorRGB(0,iVar1,iVar1,1,1);
-    DungeonRunFrameActions(0x46);
-  }
-  sub_803E708(4,0x46);
-  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 10)
-  {
-    SetDungeonBGColorRGB(iVar1 / 2,iVar1,iVar1,1,1);
-    DungeonRunFrameActions(0x46);
-  }
-  sub_803E708(4,0x46);
-  sub_8085EB0();
-}
-
-void sub_8088574(void)
-{
-  int iVar1;
-
-  for(iVar1 = 0; iVar1 <= 80; iVar1 += 8)
-  {
-    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,0,0);
-    DungeonRunFrameActions(0x46);
-  }
-}
-
-void sub_80885A0(void)
-{
-  DungeonFadeOutBGM(0x3c);
-  sub_803E708(0x3c,0x46);
-  DungeonStopBGM();
-  gDungeon->unk7 = 1;
-}
-
-void sub_80885C4(void)
-{
-  Entity * leaderEntity;
-
-  leaderEntity = xxx_call_GetLeader();
-  DungeonFadeOutBGM(0x3c);
-  sub_803E708(0x3c,0x46);
-  DungeonStopBGM();
-  sub_80854D4();
-  sub_8085930(DIRECTION_NORTH);
-  sub_80855E4(sub_8086A3C);
-  sub_8085860(leaderEntity->pos.x,leaderEntity->pos.y - 3);
-}
-
-void sub_8088608(void)
-{
-    gDungeon->unk2 = 1;
-}
-
-void sub_8088618(void)
-{
-    sub_8086448();
-    // The mountain's summit...
-    // But there is no one here
-    // It's time to go back
-    sub_80866C4(&gUnknown_8102B10);
-}
+//
 
 void sub_808862C(void)
 {
@@ -1044,6 +469,8 @@ void GroudonScreenFlash2(void)
   sub_803E708(10,0x46);
   sub_8085EB0();
 }
+
+// FILE SPLIT
 
 
 void sub_8088DC0(void)
@@ -4079,7 +3506,4 @@ void sub_808CD9C(void)
     sub_80866C4(&gUnknown_810697C);
 }
 
-const u64 *sub_808CDB0(u8 r0)
-{
-    return &gUnknown_8107544[r0];
-}
+
